+++ cmd/main.go
@@ -14,6 +14,7 @@ import (
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/server"
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/auth"
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/category"
+	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/citizencard"
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/media"
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/message"
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/object"
@@ -22,7 +23,6 @@ import (
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/quotation"
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/review"
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/user"
-	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/verify"
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/validator"
 	"github.com/CP-RektMart/pic-me-pls-backend/pkg/logger"
 )
@@ -60,7 +60,7 @@ func main() {
 	authHandler := auth.NewHandler(store, validate, jwtService, authMiddleware, config.GoogleClientID)
 	userHandler := user.NewHandler(store, validate, authMiddleware)
 	photographersHandler := photographers.NewHandler(store, validate, authMiddleware)
-	verifyHandler := verify.NewHandler(store, validate, authMiddleware)
+	citizencardHandler := citizencard.NewHandler(store, validate, authMiddleware)
 	packageHandler := packages.NewHandler(store, validate, authMiddleware)
 	reviewHandler := review.NewHandler(store, validate)
 	categoryHandler := category.NewHandler(store, validate)
@@ -77,7 +77,7 @@ func main() {
 		authHandler,
 		userHandler,
 		photographersHandler,
-		verifyHandler,
+		citizencardHandler,
 		packageHandler,
 		reviewHandler,
 		categoryHandler,
