+++ internal/services/citizencard/get-citizencard.go
@@ -1,4 +1,4 @@
-package verify
+package citizencard
 
 import (
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/dto"
