+++ internal/services/citizencard/reverify-card.go
@@ -1,4 +1,4 @@
-package verify
+package citizencard
 
 import (
 	"mime/multipart"
