+++ internal/services/citizencard/verify-card.go
@@ -1,4 +1,4 @@
-package verify
+package citizencard
 
 import (
 	"context"
